+++ src/msw/frame.cpp
@@ -56,6 +56,11 @@
     #include "wx/univ/colschem.h"
 #endif // __WXUNIVERSAL__
 
+#if wxUSE_TASKBARBUTTON
+    #include "wx/taskbarbutton.h"
+#endif
+
+
 // ----------------------------------------------------------------------------
 // globals
 // ----------------------------------------------------------------------------
@@ -65,6 +70,7 @@
 #endif // wxUSE_MENUS || wxUSE_MENUS_NATIVE
 
 #if wxUSE_TASKBARBUTTON
+    static WXUINT gs_msgTaskbarButtonCreated = 0;
     #define wxTHBN_CLICKED 0x1800
 #endif  // wxUSE_TASKBARBUTTON
 
@@ -136,7 +142,25 @@ bool wxFrame::Create(wxWindow *parent,
     SetAcceleratorTable(accel);
 #endif // wxUSE_ACCEL && __POCKETPC__
 
+#if wxUSE_TASKBARBUTTON
+    m_taskBarButton = NULL;
+    gs_msgTaskbarButtonCreated =
+        ::RegisterWindowMessage(wxT("TaskbarButtonCreated"));
+#endif
+
     return true;
+}
+
+wxFrame::~wxFrame()
+{
+    SendDestroyEvent();
+
+    DeleteAllBars();
+
+#if wxUSE_TASKBARBUTTON
+    if ( m_taskBarButton )
+        delete m_taskBarButton;
+#endif
 }
 
 // ----------------------------------------------------------------------------
@@ -473,6 +497,13 @@ wxMenu* wxFrame::MSWFindMenuFromHMENU(WXHMENU hMenu)
     return GetMenuBar() ? GetMenuBar()->MSWGetMenu(hMenu) : NULL;
 }
 #endif // wxUSE_MENUS && !defined(__WXUNIVERSAL__)
+
+#if wxUSE_TASKBARBUTTON
+wxTaskBarButton* wxFrame::MSWGetTaskBarButton()
+{
+    return m_taskBarButton;
+}
+#endif // wxUSE_TASKBARBUTTON
 
 // Responds to colour changes, and passes event on to children.
 void wxFrame::OnSysColourChanged(wxSysColourChangedEvent& event)
@@ -867,9 +898,11 @@ bool wxFrame::HandleCommand(WXWORD id, WXWORD cmd, WXHWND control)
 #endif // wxUSE_MENUS
 
 #if wxUSE_TASKBARBUTTON
-    if ( cmd == wxTHBN_CLICKED && MSWGetTaskBarButton() )
+    if ( cmd == wxTHBN_CLICKED && m_taskBarButton )
     {
-        return wxTopLevelWindowMSW::HandleTHBNClickedCommand(id);
+        wxCommandEvent event(wxEVT_BUTTON, id);
+        event.SetEventObject(this);
+        return ProcessEvent(event);
     }
 #endif // wxUSE_TASKBARBUTTON
 
@@ -927,6 +960,13 @@ WXLRESULT wxFrame::MSWWindowProc(WXUINT message, WXWPARAM wParam, WXLPARAM lPara
             break;
 #endif // !__WXMICROWIN__
     }
+#if wxUSE_TASKBARBUTTON
+    if ( message == gs_msgTaskbarButtonCreated )
+    {
+        m_taskBarButton = new wxTaskBarButtonImpl(GetHandle());
+        processed = true;
+    }
+#endif
 
     if ( !processed )
         rc = wxFrameBase::MSWWindowProc(message, wParam, lParam);
